requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6
Flask-JWT-Extended==4.6.0
Werkzeug==2.3.3
Requests==2.32.3
redis==4.3.4